+++ src/video_output/vout_intf.c
@@ -98,7 +98,7 @@ void *vout_RequestWindow( vout_thread_t *p_vout,
 
     /* Check whether someone provided us with a window ID */
     var_Get( p_vout->p_libvlc, "drawable", &val );
-    if( val.i_int ) return (void *)val.i_int;
+    if( val.i_int ) return (void *)(intptr_t)val.i_int;
 
     /* Find if the main interface supports embedding */
     p_list = vlc_list_find( p_vout, VLC_OBJECT_INTF, FIND_ANYWHERE );
